ambulance-dispatch/backend/services/siren-detection/requirements.txt
numpy
scipy
sounddevice
opencv-python
pytesseract
Pillow
colorama
fastapi
uvicorn
python-multipart
